+++ android/app/src/main/java/com/brekeke/phonedev/activity/IncomingCallActivity.java
@@ -1,4 +1,4 @@
-package com.brekeke.phonedev;
+package com.brekeke.phonedev.activity;
 
 import android.Manifest.permission;
 import android.app.Activity;
@@ -38,8 +38,13 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.core.content.res.ResourcesCompat;
 import androidx.swiperefreshlayout.widget.CircularProgressDrawable;
-import com.brekeke.phonedev.toast.ToastManager;
-import com.brekeke.phonedev.toast.ToastType;
+import com.brekeke.phonedev.BrekekeUtils;
+import com.brekeke.phonedev.BuildConfig;
+import com.brekeke.phonedev.MainActivity;
+import com.brekeke.phonedev.R;
+import com.brekeke.phonedev.utils.L;
+import com.brekeke.phonedev.utils.ToastManager;
+import com.brekeke.phonedev.utils.ToastType;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.load.engine.DiskCacheStrategy;
 import com.facebook.react.bridge.ReadableArray;
@@ -50,6 +55,7 @@ import java.util.Timer;
 import java.util.TimerTask;
 import org.json.JSONObject;
 
+// incoming call screen
 public class IncomingCallActivity extends Activity implements View.OnClickListener {
   private ToastManager toastManager;
   private LinearLayout toastContainer;
@@ -206,7 +212,7 @@ public class IncomingCallActivity extends Activity implements View.OnClickListen
       BrekekeUtils.staticStartRingtone();
     }
 
-    // Initialize toast manager
+    // initialize toast manager
     toastContainer = (LinearLayout) findViewById(R.id.toast_container);
     toastManager = ToastManager.getInstance(this, toastContainer);