+++ android/app/src/main/java/com/brekeke/phonedev/push_notification/ReactNativeJson.java
@@ -1,4 +1,4 @@
-package com.brekeke.phonedev;
+package com.brekeke.phonedev.push_notification;
 
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.bridge.ReadableMap;