+++ android/app/src/main/java/com/brekeke/phonedev/utils/L.java
@@ -1,5 +1,6 @@
-package com.brekeke.phonedev;
+package com.brekeke.phonedev.utils;
 
+// labels for multi langues
 public class L {
   public static String l = null;
 
@@ -87,5 +88,13 @@ public class L {
 
   public static String goToSetting() {
     return "ja".equals(l) ? "設定に移動" : "Go to settings";
+  }
+
+  public static String serviceIsRunning() {
+    return "ja".equals(l) ? "サービスは実行中です" : "Service is running";
+  }
+
+  public static String serviceIsRunningInBackground() {
+    return "ja".equals(l) ? "サービスはバックグラウンドで実行されています" : "Service is running in background";
   }
 }